requirements.txt
selenium>=4.10.0
webdriver-manager>=3.8.6
beautifulsoup4>=4.12.2
polars>=0.19.0
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0
click>=8.1.3